// frontend/src/shared/api/generated/model/slotStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type SlotStatus = (typeof SlotStatus)[keyof typeof SlotStatus];

export const SlotStatus = {
  EMPTY: 'EMPTY',
  OCCUPIED: 'OCCUPIED',
  RECOGNIZING: 'RECOGNIZING',
  RECOGNITION_FAILED: 'RECOGNITION_FAILED',
} as const;